refactor: componentize forwarder
